Tidy up item handling
